Replace LocationChangedEventListener with LocationNavigator service
